requirements.txt
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/nix/avx2
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/nix/generic
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic
appdirs==1.4.4
attrs==19.3.0
click==7.1.2
cycler==0.10.0
joblib==1.0.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.4
pathspec==0.8.0
pyparsing==2.4.5
python-dateutil==2.8.1
regex==2019.11.1
scikit-learn==0.23.0
scipy==1.3.2
six==1.13.0
sklearn==0.0
tensorflow_cpu==1.15
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
